assembly: share exported type error helper
